Refactored molecular properties into a utility class.
